Refactor: moved TreeViewProvider to a separate class
